Adds functionality to 'Rotation Input' nodes that allows the User to access the combined (implied) rotations for a pose bone, instead of previously only having access to the deliberately set by the user 'rotation_euler' or 'rotation_quaternion' pose bone properties, which ignore external rotations. This allows the user access to rotations to the bone caused by Constraints or other external sources, while also combining them to directly set 'rotation_*' properties.
This is useful, for example, when using a rotation as the RBF driver input, and taking the rotation applied by an IK constraint to a humerus bone in that IK chain, to assist in rigging a character's shoulder mesh shapekeys or helper bone transforms.

Changed the 'rotation_input' node to have a new checkbox that allows the user to use the accumulated rotations to a bone.

Edited the 'common' file to include functions that perform the math matrix transforms to accesss these relevant rotations.

Created new variable in 'strings' file: 'INCLUDE_EXTERNAL_ROTATIONS_LABEL', and introduced locale translations for that variable in english, portuguese and spanish.